# Get rid off implicit new/delete by std::vector when invoking a function