--- a/include/pybind11/functional.h
+++ b/include/pybind11/functional.h
@@ -46,11 +46,13 @@ public:
             auto c = reinterpret_borrow<capsule>(PyCFunction_GET_SELF(cfunc.ptr()));
             auto rec = (function_record *) c;
 
-            if (rec && rec->is_stateless &&
-                    same_type(typeid(function_type), *reinterpret_cast<const std::type_info *>(rec->data[1]))) {
-                struct capture { function_type f; };
-                value = ((capture *) &rec->data)->f;
-                return true;
+            if (rec)
+            {
+                if (void* ptr = rec->try_get_function_pointer(typeid(function_type)))
+                {
+                    value = reinterpret_cast<function_type>(ptr);
+                    return true;
+                }
             }
         }