Hi all, I've implemented the idea proposed in https://github.com/pybind/pybind11/issues/2005
It does significantly improve performance of a single function call, there is no new/deletes anymore. 
It has also doubled the size of the binary, but I haven't spent any time reducing template bloat. It's just a proof of concept. I'm happy to clean it up if main devs find it useful

old
588 ns ± 113 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)

new 
373 ns ± 5.99 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
------ pybind11_tests.cpython-37m-x86_64-linux-gnu.so file size: 649264

pure python
250 ns ± 9.25 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
------ pybind11_tests.cpython-37m-x86_64-linux-gnu.so file size: 3293632